puts semester switchover logic into cadence strategy
